pyproject.toml
[project]
name = "python-stripe-example"
version = "0.0.0"
description = "Example project showing how to integrate Stripe payments into Python applications"
readme = "README.md"
dependencies = [
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "stripe>=14.1.0",
]
requires-python = "==3.14.2"
